requirements.txt
spacy>=3.5.1
torch>=2.0.0
torchdata>=0.6.0
torchtext>=0.15.1
pyyaml>=6.0